requirements.txt
fastapi
uvicorn
motor
python-dotenv
httpx
python-multipart
fastapi-mail